+++ packages/webcomponentsjs/gulpfile.js
@@ -104,6 +104,20 @@ function closurify(sourceName, fileName) {
   .pipe(gulp.dest(outDir));
 }
 
+gulp.task('debugify-pf_js', () => {
+  const rollupOptions = {
+    plugins: [commonjs()]
+  };
+  return debugify('webcomponents-pf_js', null, rollupOptions)
+});
+
+gulp.task('debugify-pf_dom', () => {
+  const rollupOptions = {
+    plugins: [commonjs()]
+  };
+  return debugify('webcomponents-pf_dom', null, rollupOptions)
+});
+
 gulp.task('debugify-ce', () => {
   return debugify('webcomponents-ce')
 });
@@ -129,6 +143,14 @@ gulp.task('debugify-bundle', () => {
   };
   return debugify('webcomponents-bundle', 'webcomponents-bundle', rollupOptions);
 })
+
+gulp.task('closurify-pf_js', () => {
+  return closurify('webcomponents-pf_js')
+});
+
+gulp.task('closurify-pf_dom', () => {
+  return closurify('webcomponents-pf_dom')
+});
 
 gulp.task('closurify-ce', () => {
   return closurify('webcomponents-ce')
@@ -177,6 +199,8 @@ gulp.task('clean', () => {
 });
 
 gulp.task('debug', gulp.series([
+  'debugify-pf_js',
+  'debugify-pf_dom',
   'debugify-ce',
   'debugify-sd',
   'debugify-sd-ce',
@@ -186,6 +210,8 @@ gulp.task('debug', gulp.series([
 ]));
 
 gulp.task('closure', gulp.series([
+  'closurify-pf_js',
+  'closurify-pf_dom',
   'closurify-ce',
   'closurify-sd',
   'closurify-sd-ce',