+++ packages/shadydom/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fix HTMLCollection length incorrect issue
+  ([#378](https://github.com/webcomponents/polyfills/pull/332))
 - Add support for ChildNode APIs.
   ([#390](https://github.com/webcomponents/polyfills/pull/390))
 - Add support for select ParentNode APIs.